+++ src/strongbox/core.clj
@@ -17,7 +17,7 @@
     [http :as http]
     [logging :as logging]
     [nfo :as nfo]
-    [utils :as utils :refer [join not-empty? false-if-nil nav-map nav-map-fn delete-many-files! static-slurp expand-path if-let*]]
+    [utils :as utils :refer [join nav-map nav-map-fn delete-many-files! static-slurp expand-path if-let*]]
     [catalogue :as catalogue]
     [specs :as sp]]))
 
@@ -36,7 +36,23 @@
    :installed/needs-updating :lemonchiffon
    :installed/hovering "#e6e6e6" ;; light grey
    :search/already-installed "#99bc6b" ;; greenish
-   :hyperlink :blue})
+   :hyperlink :blue
+
+   ;; jfx
+   :base "#ececec"
+   :accent "lightsteelblue"
+   :button-text-hovering "black"
+   :table-border "#bbb"
+   :row "-fx-control-inner-background"
+   :row-hover "derive(-fx-control-inner-background,-10%)"
+   :unsteady "lightsteelblue"
+   :row-updateable "lemonchiffon"
+   :row-warning "lemonchiffon"
+   :row-error "tomato"
+   :jfx-hyperlink "blue"
+   :jfx-hyperlink-weight "normal"
+   :table-font-colour "derive(-fx-background,-80%)"
+   :already-installed-row-colour "#99bc6b"})
 
 ;; inverse colours of -colour-map
 (def -dark-colour-map
@@ -48,7 +64,24 @@
    :installed/needs-updating "#000532"
    :installed/hovering "#191919"
    :search/already-installed "#664394"
-   :hyperlink :yellow})
+   :hyperlink :yellow
+
+   ;; jfx
+   ;; https://github.com/dracula/dracula-theme
+   :base "#1e1f29"
+   :accent "#44475a"
+   :button-text-hovering "white"
+   :table-border "#333"
+   :row "#1e1f29" ;; same as :base
+   :row-hover "derive(-fx-control-inner-background,-10%)"
+   :unsteady "-fx-selection-bar"
+   :row-updateable "#6272a4"
+   :row-warning "#6272a4"
+   :row-error "#ce2828"
+   :jfx-hyperlink "#f8f8f2"
+   :jfx-hyperlink-weight "bold"
+   :table-font-colour "white"
+   :already-installed-row-colour "#99bc6b"})
 
 (def themes
   {:light -colour-map
@@ -149,6 +182,9 @@
    ;; the root swing window
    :gui nil
 
+   ;; jfx ui showing?
+   :gui-showing? false
+
    ;; set to anything other than `nil` to have `main.clj` restart the gui
    :gui-restart-flag nil
 
@@ -163,6 +199,10 @@
    :selected-installed []
 
    :search-field-input nil
+
+   ;; results of searching db for `:search-field-input`
+   :search-results []
+
    :selected-search []
    ;; number of results to display in search results pane.
    ;; adjust to whatever performs the best
@@ -206,6 +246,12 @@
    :get-etag #(get-state :etag-db %) ;; do this instead
    :cache-dir (paths :cache-dir)})
 
+(defn-spec add-cleanup-fn nil?
+  "adds a function to a list of functions that are called without arguments when the application is stopped"
+  [f fn?]
+  (swap! state update-in [:cleanup] conj f)
+  nil)
+
 (defn-spec state-bind nil?
   "executes given callback function when value at path in state map changes. 
   trigger is discarded if old and new values are identical"
@@ -225,8 +271,7 @@
                      (callback new-state)
                      (catch Exception e
                        (error e "error caught in watch! your callback *must* be catching these or the thread dies silently:" path))))))
-
-    (swap! state update-in [:cleanup] conj rmwatch)
+    (add-cleanup-fn rmwatch)
     nil))
 
 ;; addon dirs
@@ -244,6 +289,7 @@
       (swap! state update-in [:cfg :addon-dir-list] conj stub))
     nil))
 
+;; see also: strongbox.ui.cli/set-addon-dir! 
 (defn-spec set-addon-dir! nil?
   "adds a new :addon-dir to :addon-dir-list (if it doesn't already exist) and marks it as selected"
   [addon-dir ::sp/addon-dir]
@@ -303,11 +349,12 @@
      nil)))
 
 (defn-spec get-game-track (s/or :ok ::sp/game-track, :missing nil?)
+  "returns the game track for the given `addon-dir` or the currently selected addon-dir if no `addon-dir` given"
   ([]
-   (when-let [addon-dir (selected-addon-dir)]
-     (get-game-track addon-dir)))
-  ([addon-dir ::sp/addon-dir]
-   (-> addon-dir addon-dir-map :game-track)))
+   (get-game-track (selected-addon-dir)))
+  ([addon-dir (s/nilable ::sp/addon-dir)]
+   (when addon-dir
+     (-> addon-dir addon-dir-map :game-track))))
 
 ;; settings
 
@@ -363,6 +410,11 @@
   [addon]
   (swap! state update-in [:unsteady-addons] clojure.set/difference #{(:name addon)}))
 
+(defn-spec unsteady? boolean?
+  "returns `true` if given `addon` is being updated"
+  [addon-name ::sp/name]
+  (utils/in? addon-name (get-state :unsteady-addons)))
+
 (defn affects-addon-wrapper
   [wrapped-fn]
   (fn [addon & args]
@@ -373,6 +425,12 @@
         (stop-affecting-addon addon)))))
 
 ;; selecting addons
+
+(defn-spec select-addons-search* nil?
+  "sets the selected list of addons in application state for a later action"
+  [selected-addons :addon/summary-list]
+  (swap! state assoc :selected-search selected-addons)
+  nil)
 
 (defn-spec select-addons* nil?
   "sets the selected list of addons to the given `selected-addons` for bulk operations like 'update', 'delete', 'ignore', etc"
@@ -971,7 +1029,7 @@
 ;;
 
 (defn refresh
-  [& _]
+  [& _] ;; todo: remove args with swing gui
   (profile
    {:when (get-state :profile?)}
 
@@ -1059,7 +1117,11 @@
 (defn-spec clear-ignore-selected nil?
   "removes the 'ignore' flag from each of the selected addons."
   []
-  (->> (get-state) :selected-installed (map :dirname) (run! (partial addon/clear-ignore (selected-addon-dir))))
+  (->> (get-state :selected-installed)
+       (mapv addon/ungroup-addon)
+       flatten
+       (mapv :dirname)
+       (run! (partial addon/clear-ignore (selected-addon-dir))))
   (refresh))
 
 ;;
@@ -1105,16 +1167,6 @@
 
 ;; init
 
-(defn watch-for-addon-dir-change
-  "when the current addon directory changes, the list of installed addons should be re-read"
-  []
-  (state-bind [:cfg :selected-addon-dir] (fn [_] (refresh))))
-
-(defn watch-for-catalogue-change
-  "when the catalogue changes, the db should be rebuilt"
-  []
-  (state-bind [:cfg :selected-catalogue] (fn [_] (db-reload-catalogue))))
-
 (defn-spec set-paths! nil?
   []
   (swap! state assoc :paths (generate-path-map))
@@ -1141,8 +1193,6 @@
   (init-dirs)
   (prune-http-cache!)
   (load-settings! cli-opts)
-  (watch-for-addon-dir-change)
-  (watch-for-catalogue-change)
 
   state)