+++ test/strongbox/core_test.clj
@@ -1094,8 +1094,71 @@
         (is (:ignore? (first (core/get-state :installed-addon-list))))
 
         (core/clear-ignore-selected)
-        (is (= expected (first (core/get-state :installed-addon-list)))))))
+        (is (= expected (first (core/get-state :installed-addon-list))))))))
 
+(deftest clear-addon-ignore-flag--group-addons
+  (testing "an addon with an ignored group member can be 'unignored'. see issue#193"
+    (with-running-app
+      (let [install-dir (helper/install-dir)
+            addon {:name "everyotheraddon" :label "EveryOtherAddon" :version "5.6.7" :url "https://group.id/also/never/fetched"
+                   :source "curseforge" :source-id 2
+                   :-testing-zipfile (fixture-path "everyotheraddon--5-6-7.zip")}
+
+            expected {:description "group record for the fetched addon",
+                      :dirname "EveryAddon-BundledAddon",
+                      :group-addon-count 2,
+                      :group-addons [{:description "A useful addon that everyone bundles with their own.",
+                                      :dirname "EveryAddon-BundledAddon",
+                                      :group-id "https://group.id/also/never/fetched",
+
+                                      :ignore? true,
+
+                                      :installed-game-track :retail,
+                                      :installed-version "5.6.7",
+                                      :interface-version 80000,
+                                      :label "BundledAddon a.b.c",
+                                      :name "everyotheraddon",
+                                      :primary? false,
+                                      :source "curseforge",
+                                      :source-id 2}
+
+                                     {:description "Does what every addon does, just better",
+                                      :dirname "EveryOtherAddon",
+                                      :group-id "https://group.id/also/never/fetched",
+                                      :installed-game-track :retail,
+                                      :installed-version "5.6.7",
+                                      :interface-version 70000,
+                                      :label "EveryOtherAddon 5.6.7",
+                                      :name "everyotheraddon",
+                                      :primary? false,
+                                      :source "curseforge",
+                                      :source-id 2}],
+                      :group-id "https://group.id/also/never/fetched",
+                      :ignore? true,
+                      :installed-game-track :retail,
+                      :installed-version "5.6.7",
+                      :interface-version 80000,
+                      :label "fetched (group)",
+                      :name "everyotheraddon",
+                      :primary? false,
+                      :source "curseforge",
+                      :source-id 2}
+
+            target-idx 0
+            expected-2 (-> expected
+                           (dissoc :ignore?)
+                           (assoc :update? false)
+                           (update-in [:group-addons target-idx] dissoc :ignore?))]
+        (core/install-addon addon)
+        (nfo/ignore install-dir "EveryAddon-BundledAddon")
+        (core/load-installed-addons)
+        (core/select-addons)
+        (is (= expected (first (core/get-state :installed-addon-list))))
+
+        (core/clear-ignore-selected)
+        (is (= expected-2 (first (core/get-state :installed-addon-list))))))))
+
+(deftest clear-addon-ignore-flag--implicit-ignore
   (testing "an implicitly ignored (vcs) addon can be 'unignored' as well"
     (with-running-app
       (let [install-dir (helper/install-dir)